pyproject.toml
[project]
name = "person-base"
version = "1.0.0"
description = "个性化知识问答系统 - 基于大语言模型API的智能问答系统"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.19.0",
    "langchain>=1.0.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.4.0",
    "openai>=1.0.0",
    "faiss-cpu>=1.7.4",
    "pypdf>=4.0.0",
    "docx2txt>=0.8",
    "unstructured>=0.10.0",
    "python-multipart>=0.0.6",
    "loguru>=0.7.0",
    "python-docx>=1.2.0",
    "pyinstaller>=6.17.0",
    "markdown>=3.10",
]

[project.scripts]
start-backend = "uvicorn backend.app.main:app --reload --host 0.0.0.0 --port 8000"
